+++ nata-monkey/src/main/java/com/nata/action/ActionType.java
@@ -10,4 +10,8 @@ public class ActionType {
     public static final String TAP = "tap";
     public static final String SWIPE= "swipe";
     public static final String INPUT= "input";
+    public static final String BACK= "back";
+    public static final String MENU= "menu";
+
+
 }