+++ nata-monkey/src/main/java/com/nata/action/SwipeAction.java
@@ -8,17 +8,23 @@ import com.nata.element.Element;
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
  * Update: 2016-01-13 21:04
  */
-public class SwipeAction extends Action{
+public class SwipeAction extends Action {
     private AdbDevice device;
-    private Element element = null;
-    public SwipeAction(Element element, AdbDevice device){
+    private SwipeDirection direction;
+
+    public SwipeAction(SwipeDirection direction,AdbDevice device) {
         super.setName(ActionType.SWIPE);
         this.device = device;
-        this.element = element;
+        this.direction = direction;
     }
 
     @Override
     public void fire() {
-
+        switch (direction){
+            case DOWN: device.swipeToDown();break;
+            case UP: device.swipeToUp();break;
+            case LEFT:device.swipeToLeft();break;
+            case RIGHT:device.swipeToRight();break;
+        }
     }
 }