+++ nata-monkey/src/main/java/com/nata/action/SwipeDirection.java
@@ -1,0 +1,13 @@
+package com.nata.action;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-03-09 23:00
+ */
+public enum SwipeDirection {
+    LEFT,
+    RIGHT,
+    UP ,
+    DOWN
+}